// src/component/Header/Header.type.ts
export interface HeaderComponentProps extends NavigationAbstractComponentProps {
    cartTotals: CartTotals;
    compareTotals: number;
    Loading: boolean;

    onBackButtonClick: (e: MouseEvent) => void;
    onCloseButtonClick: (e: MouseEvent) => void;
    onSearchBarFocus: () => void;
    onClearSearchButtonClick: () => void;
    onMyAccountButtonClick: () => void;
    onSearchBarChange: (e: ChangeEvent<HTMLInputElement> | { target: { value: string } }) => void;
    isWishlistLoading: boolean;
    onEditButtonClick: (e: MouseEvent) => void;
    onMinicartButtonClick: () => void;
    onOkButtonClick: (e: MouseEvent) => void;
    onCancelButtonClick: () => void;
    onSearchOutsideClick: () => void;
    onMyAccountOutsideClick: () => void;
    onMinicartOutsideClick: () => void;
    isClearEnabled: boolean;
    searchCriteria: string;
    shareWishlist: () => void;
    header_logo_src: string;
    logo_alt: string;
    logo_height: number;
    logo_width: number;
    isLoading: boolean;
    showMyAccountLogin: boolean;
    isCheckout: boolean;
    onSignIn: () => void;
    hideActiveOverlay: () => void;
    device: Device;
    firstname?: string;
    shouldRenderCartOverlay?: boolean;
    activeOverlay: string;
    productsInWishlist: Record<string, IndexedWishlistProduct>;
    isMenuVisible: boolean;
}